+++ src/rfc4315/mod.rs
@@ -3,3 +3,4 @@
 //! <https://www.rfc-editor.org/rfc/rfc4315>
 
 pub mod appenduid;
+pub mod expunge_uid;